requirements.txt
pynacl
typer
pystache